fix(release): resolve build issues caused by export handling and C++ stride computation
- (inknit) Fixed a multiple definition error in release builds caused by
  functions marked as `static INKNIT_ALWAYS_INLINE` being defined in
  `proto_internal.h.in`. These functions are exportable for benchmarking in
  debug builds, but improperly included in release builds as well.
- (inknit_cxx) Fixed a C++ build error in `inknit::details::stride` template
  function where passing `std::uint16_t` led to type deduction failures due
  to intermediate promotion to `int`. Now uses `std::uint32_t` during
  computation and casts the result back to `std::uint16_t` as needed.

+++ src/inknit_cxx/inc/inknit.hpp
@@ -53,7 +53,9 @@ public:
 			  details::align(INKNIT_TARGET_BIT_WIDTH),
 			  width,
 			  height,
-			  details::stride<std::uint16_t>(width, bpp, INKNIT_TARGET_BIT_WIDTH),
+			  static_cast<std::uint16_t>(
+				  details::stride<std::uint32_t>(width, bpp, INKNIT_TARGET_BIT_WIDTH)
+			  ),
 			  nullptr
 		  ) {
 		std::uint32_t *const data = buffer_.data();
